// superset-frontend/plugins/plugin-chart-custom-label-bar/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import {
  QueryFormColumn,
  QueryFormData,
  QueryFormMetric,
  SetDataMaskHook,
} from './adapters/supersetAdapter';

export interface CustomLabelBarStylesProps {
  height: number;
  width: number;
}

export interface CustomLabelBarCustomizeProps {
  groupby: QueryFormColumn[];
  metric: QueryFormMetric;
  row_limit?: number;
  barColor?: string | null;
  barHeight?: number | string;
  numberFormat?: string;
  valueSuffix?: string;
}

export type CustomLabelBarFormData = QueryFormData &
  CustomLabelBarCustomizeProps;

export interface BarDatum {
  key: string;
  label: string;
  value: number;
  formattedValue: string;
  ratio: number;
}

export type CustomLabelBarProps = CustomLabelBarStylesProps & {
  bars: BarDatum[];
  barColor: string | null;
  barHeight: number;
  groupbyLabel: string;
  metricLabel: string;
  emitCrossFilters: boolean;
  selectedValues: string[];
  setDataMask: SetDataMaskHook;
};